backend/requirements.txt
fastapi
uvicorn[standard]
sqlite-utils
pydantic>=2.0
passlib[bcrypt]
python-multipart
python-jose[cryptography]
cryptography
groq
chromadb
requests
beautifulsoup4
lxml
pydantic-settings
python-dotenv
httpx
email-validator

